Change dtype of self.nei_type to np.int32.

+++ deepmd/descriptor/se_a.py
@@ -166,7 +166,7 @@ class DescrptSeA (DescrptSe):
         self.place_holders = {}
         self.nei_type = np.array([])
         for ii in range(self.ntypes):
-            self.nei_type = np.append(self.nei_type, ii * np.ones(self.sel_a[ii])) # like a mask
+            self.nei_type = np.append(self.nei_type, ii * np.ones(self.sel_a[ii], dtype=np.int32)) # like a mask
 
         avg_zero = np.zeros([self.ntypes,self.ndescrpt]).astype(GLOBAL_NP_FLOAT_PRECISION)
         std_ones = np.ones ([self.ntypes,self.ndescrpt]).astype(GLOBAL_NP_FLOAT_PRECISION)
